Split up MessageViews into multiple files
