--- AuroraEditor/Base/AuroraEditorTextView/Model/Shims.swift
@@ -1,13 +1,0 @@
-//
-//  Shims.swift
-//  Pods
-//
-//  Created by Kabir Oberai on 19/06/18.
-//
-//
-
-import Foundation
-import AppKit
-
-/// AttributedStringKey
-public typealias AttributedStringKey = NSAttributedString.Key
